backend/pyproject.toml
[project]
name = "hype-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "google-genai>=1.70.0",
    "httpx>=0.28.1",
    "msoffcrypto-tool>=6.0.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pydantic>=2.12.5",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.24",
    "python-telegram-bot>=22.7",
    "requests>=2.33.1",
    "supabase>=2.28.3",
    "uvicorn[standard]>=0.43.0",
]
